project plan button add commonly in setup page
