+++ src/App.scss
@@ -278,7 +278,6 @@ fieldset:disabled .btn {
       // border-top: 1px solid #acacac !important;
       font-size: 12px;
       // line-height: 20px;
-      border-top: 0.5px solid #acacac !important;
       color: #595959;
       // text-align: center;
       background: #ffffff !important;
