+++ src/components/ProjectSetup/index.js
@@ -1,41 +1,325 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React, { useEffect, useState, useRef } from "react";
+import moment from "moment";
+import { Toast } from "primereact/toast";
+import { isArray } from "lodash";
 import AddProject from "../Projects/CreateProject";
 import { DropdownButton, Dropdown, Accordion } from "react-bootstrap";
 import "primeicons/primeicons.css";
 import "./index.scss";
-import { useSelector } from "react-redux";
+import { useSelector, useDispatch } from "react-redux";
 import ProjectPlanCompo from "../Projects/ProjectPlan/ProjectPlanCompo";
 import ConfirmationDialog from "./confirmationDialog";
 import TabsComponent from "./tabsComponent";
 import { hasAllAccess } from "../../utils";
+import { getProjectPlan } from "../../apis/projectPlanApi";
+import {
+  updateProjectPlanAction,
+  updateProjectPlanDesignAction,
+} from "../../store/actions/ProjectPlanActions";
+import {
+  activateProjectPlan,
+  saveProjectPlanAction,
+} from "../../apis/projectPlanApi";
+import { getMyProject } from "../../store/actions/ProjectActions";
 import { NavLink, useLocation, useNavigate, useParams } from "react-router-dom";
+import { Button } from "react-bootstrap";
 import ProjectListHeader from "../Projects/MyProjects/ProjectListHeader";
 import { ProjectService } from "../../service/PegaService";
 import ArtworkAlignment from "../DSBP/ArtworkAlignmentPage";
+
+
 function ProjectSetup(props) {
+  const toast = useRef(null);
+  const isAccessEmpty = hasAllAccess();
+  
   const projectSetup = useSelector((state) => state.ProjectSetupReducer);
   const selectedProjectDetails = projectSetup.selectedProject;
   const { mode } = projectSetup;
   const User = useSelector((state) => state.UserReducer);
   const userInformation = User.userInformation;
   const { accessMatrix } = useSelector((state) => state?.accessMatrixReducer);
+  const { projectPlan, projectPlanDesign, loading } = useSelector((state) => state.ProjectPlanReducer);
+  const { myProject, ...myProjectData } = useSelector(
+    (state) => state.myProject
+  );
+
+  const columnNames = ProjectService.getProjectPlanAllColumnNames();
   let { ProjectID } = useParams();
   let path = "";
   if (window?.location?.pathname.includes("projectPlan")) {
     path = "/projectPlan";
   }
   const navigate = useNavigate();
+  const location = useLocation();
+  const dispatch = useDispatch();
+  const locationPath = location?.pathname;
+  const url = locationPath?.split("/");
   const [toggleButtons, setToggleButtons] = useState("Tabular");
   const [option, setOption] = useState("");
   const [visible, setVisible] = useState(false);
-  const location = useLocation();
-  const locationPath = location?.pathname;
-  const url = locationPath?.split("/");
+  const [activeSave, setActiveSave] = useState(true);
+  const [tabName, setTabName] = useState(url[2] !== undefined ? url[2] : url[1]);
+  const [tabNameForPP, setTabNameForPP] = useState("Design");
+  const [projectPlanDesignData, setProjectPlanDesignData] = useState([]);
+  const [updatedProjectPlanDesignData, setUpdatedProjectPlanDesignData] =
+    useState([]);
+    const [loader, setLoader] = useState(false);
+    const [activeFlag, setActiveFlag] = useState(isAccessEmpty);
+  
+  
   const getData = (option) => {
     setVisible(true);
     setOption(option);
   };
+  useEffect(() => {
+    if (!isAccessEmpty) {
+      setActiveSave(true);
+    }
+  }, [isAccessEmpty]);
+
+  useEffect(() => {
+    setActiveFlag(false);
+    let projectData = isArray(myProject) && myProject.find(
+      (project) => project.Project_ID === ProjectID
+    );
+    const firstTime = projectPlanDesign.some(
+      (item) => item.Assignee !== "" || item.Role !== ""
+    );
+    if (
+      (!firstTime && projectData?.Project_State === "Draft") ||
+      projectData?.Project_State === "Active" ||
+      !isAccessEmpty ||
+      projectPlan.length === 0
+    ) {
+      setActiveFlag(true);
+    }
+  }, [myProject, projectPlan, isAccessEmpty, projectPlanDesign]);
+
+  useEffect(() => {
+    if (updatedProjectPlanDesignData) {
+      setProjectPlanDesignData(updatedProjectPlanDesignData || []);
+    }
+  }, [updatedProjectPlanDesignData]);
+
+  useEffect(() => {
+    getProjectPlanApi();
+  }, [mode, tabNameForPP]);
+
+  const getProjectPlanApi = async () => {
+    setLoader(true);
+    let restructuredData = [];
+    const apiData =
+      (mode === "edit" || mode === "design") && selectedProjectDetails.Project_ID
+        ? await getProjectPlan(selectedProjectDetails.Project_ID)
+        : [];
+    apiData && dispatch(updateProjectPlanDesignAction(apiData));
+    restructuredData = apiData?.length > 0 ? getRestructuredData(apiData) : [];
+    dispatch(updateProjectPlanAction(restructuredData));
+    setLoader(false);
+  };
+
+  const getRestructuredData = (apiData) => {
+    let mainTempArr = [];
+    let tasks = [];
+    if (tabNameForPP === "Input") {
+      tasks = [
+        {
+          name: "Start Artwork Alignment",
+          code: "SAA",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("SAA_")),
+        },
+      ];
+    } else {
+      tasks = [
+        {
+          name: "Define Design Intent",
+          code: "DDI",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("DDI_")),
+        },
+        {
+          name: "Upload Approved Design Intent",
+          code: "UADI",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("UADI_")),
+        },
+        {
+          name: "Define Design Template",
+          code: "DDT",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("DDT_")),
+        },
+        {
+          name: "Upload Regional Design Template",
+          code: "URDT",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("URDT_")),
+        },
+        {
+          name: "Approve Regional Design Template",
+          code: "ARDT",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("ARDT_")),
+        },
+        {
+          name: "Define Production Ready Art",
+          code: "DPRA",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("DPRA_")),
+        },
+        {
+          name: "Upload Production Ready Art",
+          code: "UPRA",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("UPRA_")),
+        },
+        {
+          name: "Approve Production Ready Art",
+          code: "APRA",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("APRA_")),
+        },
+
+        {
+          name: "Confirm Preliminary print feasibility Assessment",
+          code: "CPPFA",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("CPPFA_")),
+        },
+        {
+          name: "Define New Print Feasibility Scope",
+          code: "DNPF",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("DNPF_")),
+        },
+        {
+          name: "Confirm Color Development",
+          code: "CCD",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("CCD_")),
+        },
+        {
+          name: "Confirm Print Trial",
+          code: "CPT",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("CPT_")),
+        },
+        {
+          name: "Define New Ink Qualification scope",
+          code: "DNIQ",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("DNIQ_")),
+        },
+        {
+          name: "Confirm New Ink Qualification",
+          code: "CNIQ",
+          data: apiData.filter((data) => data.AWM_Task_ID.includes("CNIQ_")),
+        },
+      ];
+    }
+
+    tasks.forEach((task) => {
+      if (task.data?.length > 0) {
+        if (
+          task.data[0]?.AWM_Task_ID.includes("DDI_") ||
+          task.data[0]?.AWM_Task_ID.includes("DDT_") ||
+          task.data[0]?.AWM_Task_ID.includes("DPRA_") ||
+          task.data[0]?.AWM_Task_ID.includes("CPPFA_") ||
+          task.data[0]?.AWM_Task_ID.includes("DNPF_") ||
+          task.data[0]?.AWM_Task_ID.includes("DNIQ_") ||
+          task.data[0]?.AWM_Task_ID.includes("SAA_")
+        ) {
+          let tempObj = {};
+          tempObj["key"] = task.data[0]?.AWM_Task_ID;
+
+          let dataObj = {};
+          dataObj["Task"] = task.data[0]?.AWM_Task_ID.includes("DNPF_")
+            ? "Define New Print Feasibility Scope"
+            : task.data[0]?.Task_Name;
+          dataObj["Dependency"] = task.data[0]?.Dependency;
+          dataObj["Role"] = task.data[0]?.Role;
+          dataObj["RoleOptions"] = task.data[0]?.RoleOptions;
+          dataObj["Assignee"] = task.data[0]?.Assignee;
+          dataObj["OwnerOptions"] = task.data[0]?.OwnerOptions;
+          dataObj["State"] = task.data[0]?.State;
+          dataObj["Duration"] = task.data[0]?.Duration;
+          dataObj["StartDate"] = task.data[0]?.Start_Date;
+          dataObj["EndDate"] = task.data[0]?.End_Date;
+          dataObj["ConsumedBuffer"] = task.data[0]?.Consumed_Buffer;
+          dataObj["HelpNeeded"] = task.data[0]?.Help_Needed;
+
+          tempObj["data"] = dataObj;
+          tempObj["children"] = [];
+          tempObj["redirect"] = true;
+
+          mainTempArr.push(tempObj);
+        } else {
+          let tempObj = {};
+          let tempArr = [];
+          let pStartDate = "";
+          let pEndDate = "";
+          let startDateArr = [];
+          let endDateArr = [];
+
+          tempObj["key"] = task.code;
+          let dataObj = {};
+          dataObj["Task"] = `${task.name} (X${task.data?.length})`;
+          dataObj["Dependency"] = task.data[0]?.Dependency;
+          dataObj["Role"] = "";
+          dataObj["RoleOptions"] = "";
+          dataObj["Assignee"] = "";
+          dataObj["OwnerOptions"] = "";
+          dataObj["State"] = "";
+          dataObj["Duration"] = "";
+
+          dataObj["ConsumedBuffer"] = "";
+          dataObj["HelpNeeded"] = false;
+
+          tempObj["data"] = dataObj;
+          tempObj["redirect"] = true;
+
+          //child array creation
+
+          task.data.forEach((dt, index) => {
+            dt.Start_Date && startDateArr?.push(dt.Start_Date);
+            dt.End_Date && endDateArr?.push(dt.End_Date);
+            pStartDate =
+              startDateArr.length > 0 &&
+              moment.min(
+                startDateArr.map((date) =>
+                  moment(date, "YYYYMMDDTHHmmss.SSS [GMT]")
+                )
+              );
+            pEndDate =
+              endDateArr.length > 0 &&
+              moment.max(
+                endDateArr.map((date) =>
+                  moment(date, "YYYYMMDDTHHmmss.SSS [GMT]")
+                )
+              );
+            let tempObj = {};
+            tempObj["key"] = dt.AWM_Task_ID;
+
+            let dataObj = {};
+            dataObj["Task"] = `${index + 1}). ${dt.Task_Name}`;
+            dataObj["Dependency"] = dt.Dependency;
+            dataObj["Role"] = dt.Role;
+            dataObj["RoleOptions"] = dt.RoleOptions;
+            dataObj["Assignee"] = dt.Assignee;
+            dataObj["OwnerOptions"] = dt.OwnerOptions;
+            dataObj["State"] = dt.State;
+            dataObj["Duration"] = dt.Duration;
+            dataObj["StartDate"] = dt.Start_Date;
+            dataObj["EndDate"] = dt.End_Date;
+            dataObj["ConsumedBuffer"] = dt.Consumed_Buffer;
+            dataObj["HelpNeeded"] = dt.Help_Needed;
+
+            tempObj["data"] = dataObj;
+            tempObj["children"] = [];
+
+            tempArr.push(tempObj);
+          });
+          dataObj["StartDate"] = pStartDate;
+          dataObj["EndDate"] = pEndDate;
+
+          tempObj["children"] = tempArr;
+
+          mainTempArr.push(tempObj);
+        }
+      }
+    });
+
+    return mainTempArr; //toBeReplacedWithapiData;
+  };
+
   const accept = () => {
     return navigate("/myProjects");
   };
@@ -43,9 +327,6 @@ function ProjectSetup(props) {
   const reject = () => {
     console.log("reject");
   };
- 
-  const [tabName, setTabName] = useState(url[2] !== undefined ? url[2] : url[1]);
-  const [tabNameForPP, setTabNameForPP] = useState("Design");
 
   let items = "";
   if (tabName === "projectSetup") {
@@ -108,8 +389,7 @@ function ProjectSetup(props) {
     // isReorderedColumn || isFilterEnabled ||
     isColWidthSet;
 
-  const { projectPlan } = useSelector((state) => state.ProjectPlanReducer);
-  const columnNames = ProjectService.getProjectPlanAllColumnNames();
+  
 
   const childFunc = useRef(null);
 
@@ -117,6 +397,56 @@ function ProjectSetup(props) {
     setIsFilterEnabled(e);
   };
 
+  const onSave = async () => {
+    let updatedData = [];
+    projectPlanDesignData.filter(
+      ({ AWM_Project_ID, AWM_Task_ID, Assignee, Role, Duration }) =>
+        projectPlanDesign.some((object2) => {
+          if (
+            AWM_Task_ID === object2.AWM_Task_ID &&
+            ((Role !== undefined && Role !== "" && Role !== object2.Role) ||
+              (Assignee !== undefined &&
+                Assignee !== "" &&
+                Assignee !== object2.Assignee) ||
+              (Duration !== undefined &&
+                Duration !== "" &&
+                Duration !== object2.Duration))
+          ) {
+            updatedData.push({
+              AWM_Project_ID: AWM_Project_ID,
+              AWM_Task_ID: AWM_Task_ID,
+              Assignee: Assignee,
+              Role: Role,
+              Duration: Duration,
+            });
+            return console.log("updatedData", updatedData);
+          }
+        })
+    );
+    if (updatedData.length !== 0) {
+      const formData = {
+        ArtworkAgilityProjects: updatedData,
+      };
+      dispatch(updateProjectPlanDesignAction(updatedProjectPlanDesignData));
+      await saveProjectPlanAction(formData, selectedProjectDetails.Project_ID);
+      setActiveSave(true);
+    }
+  };
+
+  const activate = async () => {
+    setLoader(true);
+    await activateProjectPlan(selectedProjectDetails.Project_ID);
+    await dispatch(getMyProject(userInformation));
+    getProjectPlanApi();
+    setLoader(false);
+    await toast.current.show({
+      severity: "success",
+      summary: "Success",
+      detail: "Project activated successfully!",
+      life: 5000,
+    });
+  };
+  
   const itemsData = [
     {
       name: "projectSetup",
@@ -207,6 +537,12 @@ function ProjectSetup(props) {
                       tabNameForPP={tabNameForPP}
                       view={toggleButtons}
                       setTabName={setTabName}
+                      updatedProjectPlanDesignData={updatedProjectPlanDesignData}
+                      setUpdatedProjectPlanDesignData={setUpdatedProjectPlanDesignData}
+                      activeSave={activeSave}
+                      setActiveSave={setActiveSave}
+                      getProjectPlanApi={getProjectPlanApi}
+                      loader={loader}
                     />
                   </Accordion.Body>
                 </Accordion.Item>
@@ -221,6 +557,12 @@ function ProjectSetup(props) {
                     tabNameForPP={tabNameForPP}
                     view={toggleButtons}
                     setTabName={setTabName}
+                    updatedProjectPlanDesignData={updatedProjectPlanDesignData}
+                    setUpdatedProjectPlanDesignData={setUpdatedProjectPlanDesignData}
+                    activeSave={activeSave}
+                    setActiveSave={setActiveSave}
+                    getProjectPlanApi={getProjectPlanApi}
+                    loader={loader}
                   />
                   </Accordion.Body>
                 </Accordion.Item>
@@ -229,105 +571,40 @@ function ProjectSetup(props) {
                   <Accordion.Body>FA Assembly</Accordion.Body>
                 </Accordion.Item>
               </Accordion>
-              {/* <nav className="subNav">
-                <div className="nav nav-tabs" id="nav-tab" role="tablist">
-                  <button
-                    className={`nav-link ${
-                      tabNameForPP === "Design" ? "active" : ""
-                    }`}
-                    onClick={() => setTabNameForPP("Design")}
-                    id="nav-design-tab"
-                    data-bs-toggle="tab"
-                    data-bs-target="#nav-design"
-                    type="button"
-                    role="tab"
-                    aria-controls="nav-design"
-                    aria-selected="true"
+              {(
+                <div className="form-buttons" style={{ background: "#FAFAFA" }}>
+                  <Button
+                    className={
+                      !isAccessEmpty ? "btn btn-disabled" : "button-layout"
+                    }
+                    variant="secondary"
+                    onClick={() => navigate("/myProjects")}
+                    disabled={!isAccessEmpty}
                   >
-                    Design
-                  </button>
-                  <button
-                    className={`nav-link ${
-                      tabNameForPP === "Input" ? "active" : ""
-                    }`}
-                    onClick={() => setTabNameForPP("Input")}
-                    id="nav-input-tab"
-                    data-bs-toggle="tab"
-                    data-bs-target="#nav-input"
-                    type="button"
-                    role="tab"
-                    aria-controls="nav-input"
-                    aria-selected="true"
+                    Cancel
+                  </Button>
+
+                  <Button
+                    className={activeSave ? "btn btn-disabled" : "button-layout"}
+                    variant="secondary"
+                    onClick={onSave}
+                    disabled={activeSave}
                   >
-                    Input
-                  </button>
-                  <button
-                    className={`nav-link ${
-                      tabNameForPP === "FAAssembly" ? "active" : ""
-                    }`}
-                    onClick={() => setTabNameForPP("FAAssembly")}
-                    id="nav-fAAssembly-tab"
-                    data-bs-toggle="tab"
-                    data-bs-target="#nav-fAAssembly"
-                    type="button"
-                    role="tab"
-                    aria-controls="nav-fAAssembly"
-                    aria-selected="false"
+                    setup Save
+                  </Button>
+
+                  <Button
+                    className="button-layout"
+                    variant="primary"
+                    onClick={activate}
+                    disabled={activeFlag}
                   >
-                    FA Assembly
-                  </button>
+                    Activate
+                  </Button>
                 </div>
-              </nav> */}
+              )}
             </div>
           )}
-          {/* <div className="tab-content" id="nav-tabContent">
-            <div
-              className={`tab-pane fade ${
-                tabNameForPP === "Design" ? "show active" : ""
-              }`}
-              id="nav-design"
-              role="tabpanel"
-              aria-labelledby="nav-design-tab"
-            >
-              <ProjectPlanCompo
-                isSearch={isSearch}
-                setColWidth={setColWidth}
-                childFunc={childFunc}
-                test={test}
-                tabNameForPP={tabNameForPP}
-                view={toggleButtons}
-                setTabName={setTabName}
-              />
-            </div>
-            <div
-              className={`tab-pane fade ${
-                tabNameForPP === "Input" ? "show active" : ""
-              }`}
-              id="nav-input"
-              role="tabpanel"
-              aria-labelledby="nav-input-tab"
-            >
-              <ProjectPlanCompo
-                isSearch={isSearch}
-                setColWidth={setColWidth}
-                childFunc={childFunc}
-                test={test}
-                tabNameForPP={tabNameForPP}
-                view={toggleButtons}
-                setTabName={setTabName}
-              />
-            </div>
-            <div
-              className={`tab-pane fade ${
-                tabNameForPP === "FAAssembly" ? "show active" : ""
-              }`}
-              id="nav-fAAssembly"
-              role="tabpanel"
-              aria-labelledby="nav-fAAssembly-tab"
-            >
-              FA Assembly Data
-            </div>
-          </div> */}
         </div>
       ),
     },
@@ -377,8 +654,10 @@ function ProjectSetup(props) {
     </DropdownButton>
   );
 
+
   return (
     <div className="content-layout">
+      <Toast ref={toast} />
       {visible && (
         <ConfirmationDialog
           visible={visible}
