+++ src/components/ProjectSetup/index.scss
@@ -196,13 +196,13 @@ $width: 100%;
     }
   }
   .form-buttons{
-    padding: 1rem 11px;
+    padding: 0 16px;
     position: absolute;
     background: white;
     text-align: right;
     width: 100%;
     left: 0;
-    bottom: 55px;
+    bottom: 0;
   }
 }
 
@@ -271,7 +271,7 @@ $width: 100%;
 
 // TabsComponent Styles Start
 .tab-content {
-  height: calc(100% - 55px);
+  height: calc(100% - 40px);
   overflow: hidden;
   position: relative;
   padding-bottom: 20px;
