+++ src/components/Projects/ProjectPlan/index.scss
@@ -80,11 +80,11 @@ $fontWeight400: 400;
   }
 
   .textAlignTreeTable {
-    height: calc(100vh - 260px);
+    height: calc(100vh - 307px);
     overflow: auto;
 
     .p-treetable-wrapper {
-      height: calc(100% - 52px);
+      height: calc(100% - 5px);
       overflow-x: auto;
   
       thead.p-treetable-thead {
